Consolidated published guides pages for ICA and SRG.
There is even more consolidation to be had for ICA and SRG
published guide pages. In shared/_published_pages.html.erb, there
are if statements that switch on the controller. If we made the
controllers consistent, these can go away.

Removed the custom JavaScript used in ICA/SRG views in favor of
Bootstrap JavaScript.

And more Bootstrappy CSS goodness!